refactor(form): address review nits + remove dead codebook modal
- ExpandableNameCell: i18n the chevron aria-label (common.ui.expand/collapse)
- CatalogueFilter.form: guard URL category JSON.parse against malformed values
- Delete unused CodebookTreeModalGraphql.tsx (last V1 PandaTable consumer);
  extract shared Codebooktree type to codebookTree.types.ts, repoint importers
