+++ src/components/form/shared/codebookTree.types.ts
@@ -1,0 +1,7 @@
+export type Codebooktree = {
+    name: string
+    uid: string
+    code?: string
+    children?: Codebooktree[]
+    isExpandable?: boolean
+}
